Refactor lights toy modules
