+++ lights.html
@@ -101,6 +101,6 @@
     <!-- WebGL Canvas -->
     <canvas id="toy-canvas" class="toy-canvas"></canvas>
 
-    <script type="module" src="assets/js/app.js"></script>
+    <script type="module" src="assets/js/toys/lights/index.ts"></script>
   </body>
 </html>
